+++ src/meaps_all_in.cpp
@@ -104,7 +104,7 @@ List meaps_all_in_cpp(const IntegerVector jr_dist, const IntegerVector p_dist, c
 
         //
         for (auto k = 0; k < n_sites; ++k) {
-          liaisons[from][col_dispo[k]] += attirances[k];
+          liaisons[from][col_dispo[k]] += static_cast<double>(attirances[k]);
         }
       }  // fin des boucles sur les from
