+++ src/meaps_multishuf_origin.cpp
@@ -136,7 +136,7 @@ using namespace Rcpp;
    std::size_t NNboot = floor( Nboot / ntr );
    if(NNboot == 0) NNboot = 1;
    // Un vecteur représentant la matrice des flux groupés.
-   std::vector< std::vector<float> > liaisons(ntr, std::vector<float> (Nx));
+   std::vector< std::vector<double> > liaisons(ntr, std::vector<double> (Nx));
    for(size_t Iboot = 0; Iboot < ntr; ++Iboot) {
      std::fill( liaisons[Iboot].begin(), liaisons[Iboot].end(), 0 );
    }
